+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins
 
 ### Changed
 - Updated bundler dependancy to '~> 2.1'
+- Added option to ignore privileged bytes
 
 ## [5.0.1] - 2019-12-10
 ### Changed